make contianer definitions forward compatible
