+++ packages/loader/container-definitions/src/fluidModule.ts
@@ -1,0 +1,10 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+import { IComponent, IFluidObject } from "@fluidframework/component-core-interfaces";
+
+export interface IFluidModule {
+    fluidExport: IComponent & IFluidObject;
+}
