+++ packages/loader/container-definitions/src/loader.ts
@@ -12,9 +12,11 @@ import {
 } from "@fluidframework/protocol-definitions";
 import { IResolvedUrl } from "@fluidframework/driver-definitions";
 import { IEvent, IEventProvider } from "@fluidframework/common-definitions";
-import { IFluidCodeDetails, IFluidModule, IFluidPackage, AttachState } from "./chaincode";
 import { IDeltaManager } from "./deltas";
 import { ICriticalContainerError, ContainerWarning } from "./error";
+import { IFluidModule } from "./fluidModule";
+import { IFluidCodeDetails, IFluidPackage } from "./fluidPackage";
+import { AttachState } from "./runtime";
 
 /**
  * Code loading interface
