+++ packages/loader/container-definitions/src/tokenProvider.ts
@@ -1,0 +1,14 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export const IFluidTokenProvider: keyof IProvideFluidTokenProvider = "IFluidTokenProvider";
+
+export interface IProvideFluidTokenProvider {
+    readonly IFluidTokenProvider: IFluidTokenProvider;
+}
+
+export interface IFluidTokenProvider extends IProvideFluidTokenProvider {
+    intelligence: { [service: string]: any };
+}
